+++ tests/api/budget/test_views.py
@@ -8,7 +8,8 @@ from plugins.snt_malaria.models import (
 )
 from plugins.snt_malaria.models.budget_assumptions import BudgetAssumptions
 from plugins.snt_malaria.models.budget_settings import BudgetSettings
-from plugins.snt_malaria.models.cost_breakdown import CostUnitType, InterventionCostBreakdownLine
+from plugins.snt_malaria.models.cost_breakdown import InterventionCostBreakdownLine
+from plugins.snt_malaria.models.cost_unit_type import CostUnitType
 from plugins.snt_malaria.permissions import SNT_SCENARIO_BASIC_WRITE_PERMISSION, SNT_SCENARIO_FULL_WRITE_PERMISSION
 from plugins.snt_malaria.tests.common_base import SNTMalariaAPITestCase
