+++ tests/api/intervention_cost_breakdown_lines/common_base.py
@@ -1,5 +1,5 @@
 from plugins.snt_malaria.models import InterventionCostBreakdownLine
-from plugins.snt_malaria.models.cost_breakdown import CostUnitType
+from plugins.snt_malaria.models.cost_unit_type import CostUnitType
 from plugins.snt_malaria.permissions import SNT_SETTINGS_READ_PERMISSION, SNT_SETTINGS_WRITE_PERMISSION
 from plugins.snt_malaria.tests.common_base import SNTMalariaAPITestCase
